Prevents the loop of player from pushing another player (#34)
Added enable/disable push when attacking to config.lua (default value is false).
Prevents the character from pulling another one, continuing in the loop until it manages to pull it or get out of range.

How the bug works: When player 1 tries to pull player 2, from a distance, player 1's character only stopped when he could pull
player 2 or when out of range (assuming that player 1 kept trying to go in the opposite direction of player 2). With this changes, if player 1 makes one movement, the push stops automatically.