Fix stored regeneration condition not using updated vocation values (#27)
Regeneration condition is not updated on login. To verify: log-in, eat food, log-out, edit vocation regeneration amount/ticks, restart server / reload vocations, log back in - condition is using the persisted values, because upon logging in conditions are temporarily kept in storedConditionList which are added to player after placing him on map.