[Feature] Add imbuement system from otservbr-global and somes fixes (#150)
Added 100% functioning of the imbuement system, checking if the imbuement is aggressive and refactoring the "Game::checkImbuements" function

Added some functions that didn't exist.

Note that these changes are subject to some bugs, which will be fixed in another pull request shortly after this one. 
The purpose of this commit is just to generate diff for the total imbuement rework that will be done next.