[Enhancement] Rework of the imbuement system and conversion to c++ (#164)
NOTE: The pull request of optimization and enhancement of ItemAttributes is needed for this commit: https://github.com/opentibiabr/canary/commit/59e8aaa2c0ef6a744d23bcbd4915d1b33af39c4a. If don't use the "ItemAttributes enhancent" commit, the imbuement will crash!

Thanks to @lgrossi for the reviews made and for helping to solve several problems.
Thanks @Glatharth for provide the script for convert tables lua in items.xml: https://gist.github.com/dudantas/548fdbc3c521327c7d5320e27adba8fb

The system has been completely converted to c++
Fixed bugs and adapted the operation to be the same as global tibia
Possibility to imbue the equipped item
Added config.lua key for enable/disable the use of storages by the imbuement system (togleImbuementShrineStorage)

The imbuement of each item is now configured in items.xml in the following settings:
<attribute key="imbuementslot" value="3">
	<attribute key="elemental damage" value ="3"/>
	<attribute key="life leech" value ="3"/>
	<attribute key="mana leech" value ="3"/>
	<attribute key="critical hit" value ="3"/>
	<attribute key="elemental protection death" value ="3"/>
	<attribute key="elemental protection earth" value ="3"/>
	<attribute key="elemental protection fire" value ="3"/>
	<attribute key="elemental protection ice" value ="3"/>
	<attribute key="elemental protection energy" value ="3"/>
	<attribute key="elemental protection holy" value ="3"/>
	<attribute key="increase speed" value ="3"/>
	<attribute key="skillbost axe" value ="3"/>
	<attribute key="skillbost sword" value ="3"/>
	<attribute key="skillbost club" value ="3"/>
	<attribute key="skillbost shielding" value ="3"/>
	<attribute key="skillbost distance" value ="3"/>
	<attribute key="skillbost magic level" value ="3"/>
	<attribute key="increase capacity" value ="3"/>
</attribute>

Where the "imbuementslot" value is from 1 to 3 and the imbuement name values are also from 1 to 3, which is the imbuement level (basic, intricate or powerful), the name must be placed as above, as it is from the name that the value of each imbuement is called by ItemParse.