Fixed potion mana amount heals, remove wrong parameter from doTargetCombatMana on potions.lua (#523)
Was being passed an extra parameter in the "doTargetCombatMana" function, this function randomizes the amount of mana a potion heals. With this parameter, the minimum "6" was always passed, which is the value of the constant "COMBAT_MANADRAIN", thus making an "ultimate mana potion", for example, heal 80, 100, 130... however, it should cure between 425 and 575.