fix: loop in movement "stepin/out" and "addItem/remove" events (#742)
A potential infinite loop was discovered in specific scenarios involving teleport squares or items that teleport players to specific locations (such as when the player logs in). This pull request addresses the issue by disabling the onStepIn event during login and implementing additional security checks to prevent similar bugs from occurring in the future.